Exquisite Corpse in C#
In the game Exquisite Corpse, participants draw either a head, body, or feet of a creature. The players don’t know how their part of the body connects to the other two, until the drawing is finished and revealed.

For this project, we write a program that mimics the Exquisite Corpse game. Using methods, we'll be able to randomly combine different parts of ASCII characters so they create a new creature.